Fix int8_t overflow in nnom_conv2d_config_t; add bare-metal port guard
kernel_size, stride_size, padding_size, dilation_size were int8_t[2],
overflowing for kernel > 127 (this model uses kernel=129).  Change to
uint16_t[2] to match nnom_shape_data_t.  Without this fix the first
Conv1D output tensor inflates from {1,492,32} to {1,61948,32} (~2 MB),
causing silent allocation failure and a store fault in model_run().

Also add NNOM_BARE_METAL guard in nnom_port.h so the firmware can
suppress stdio.h/stdlib.h and printf-based logging on bare-metal targets.

+++ inc/layers/nnom_conv2d.h
@@ -55,11 +55,11 @@ typedef struct _nnom_conv2d_config_t
 	nnom_tensor_t *bias;
 	nnom_qformat_param_t *output_shift;   
 	nnom_qformat_param_t *bias_shift;   
-	uint32_t filter_size;  
-	int8_t kernel_size[2];
-	int8_t stride_size[2];
-	int8_t padding_size[2];
-	int8_t dilation_size[2];
+	uint32_t filter_size;
+	uint16_t kernel_size[2];
+	uint16_t stride_size[2];
+	uint16_t padding_size[2];
+	uint16_t dilation_size[2];
 	nnom_padding_t padding_type;
 } nnom_conv2d_config_t;
